% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfrating.R
\name{selfrating}
\alias{selfrating}
\title{Rating based on your name height and weight}
\usage{
selfrating(name, height, weight)
}
\arguments{
\item{name}{your first name}

\item{height}{in meter}

\item{weight}{in kg}
}
\value{
a rating sentence based on your input
}
\description{
Rating based on your name height and weight
}
\examples{
selfrating(Alex, 1.75, 70)
}
